+++ cpp/utils/a_star.h
@@ -21,6 +21,6 @@ struct Node
     float f;
 };
 
-extern void astar(float params[], int params_size);
+extern void astar(double params[], int params_size);
 extern float final_path[2050];
 extern int final_path_size;
